Modify my add and delete file
